ImagesService logging

+++ BooksStore/Services/ImagesService.cs
@@ -21,16 +21,30 @@ namespace BooksStore.Services
 
         public void DeleteImage(Book book)
         {
+            if (book == null)
+            {
+                throw new ArgumentNullException(nameof(book));
+            }
+
+            _logger.LogInformation($"DeleteImage book id '{book.Id}'");
+
             DeleteFile(book.Image);
             DeleteFile(book.ImagePreview);
         }
 
-        public bool SaveImage(Book book, Stream stream, string imageContentType) 
+        public bool SaveImage(Book book, Stream stream, string imageContentType)
         {
+            if (book == null)
+            {
+                throw new ArgumentNullException(nameof(book));
+            }
+
             if (stream == null)
             {
                 return false;
             }
+
+            _logger.LogInformation($"SaveImage book id '{book.Id}'");
 
             ValidateImage(stream, imageContentType, out string extension);
 
@@ -43,6 +57,8 @@ namespace BooksStore.Services
 
             var fileName = $"{book.Id}.{extension}";
             var previewFileName = $"{book.Id}-preview.{extension}";
+            var imagePath = Path.Combine(directory, fileName);
+            var previewPath = Path.Combine(directory, previewFileName);
 
             //TODO Оптимизация и превью, только если размер превышает заданный порог
             //TODO При высоких нагрузках возможны проблемы с памятью - использовать SemaphoreSlim при работе с изображениями
@@ -50,12 +66,19 @@ namespace BooksStore.Services
             using (var optimized = original.Resize(_settings.ImageOptimizedHeight))
             using (var preview = original.Resize(_settings.ImagePreviewHeigt))
             {
-                optimized.Save(Path.Combine(directory, fileName));
-                preview.Save(Path.Combine(directory, previewFileName));
+                optimized.Save(imagePath);
+                _logger.LogInformation($"Saved: {imagePath}");
+
+                preview.Save(previewPath);
+                _logger.LogInformation($"Saved: {imagePath}");
             }
 
             book.Image = $"{_settings.ImagesRelativePath}/{fileName}";
             book.ImagePreview = $"{_settings.ImagesRelativePath}/{previewFileName}";
+
+            _logger.LogInformation($"Set {nameof(Book.Image)} property for book id '{book.Id}': {book.Image}");
+            _logger.LogInformation($"Set {nameof(Book.ImagePreview)} property for book id '{book.Id}': {book.ImagePreview}");
+
 
             return true;
 
@@ -90,6 +113,7 @@ namespace BooksStore.Services
             if (File.Exists(path))
             {
                 File.Delete(path);
+                _logger.LogInformation($"Deleted: {path}");
             }
         }
     }
